+++ lib/data/models/food_cart_item.dart
@@ -2,18 +2,17 @@ import 'package:flutter_foods/data/models/food.dart';
 
 class FoodCartItem {
   final Food food;
-  int quantity =1;
-  bool isChecked; 
+  int quantity = 1;
+  bool isChecked;
 
   FoodCartItem({
     required this.food,
     required this.quantity,
-    this.isChecked = false, 
+    this.isChecked = false,
   });
-  double get totalPrice => food.price * quantity;
-FoodCartItem.copy(FoodCartItem other)
+  int get totalPrice => food.price * quantity;
+  FoodCartItem.copy(FoodCartItem other)
       : food = other.food,
         quantity = other.quantity,
         isChecked = other.isChecked;
-
 }
